refactor : Ladder 와 Line을 연결하는 Lines를 삭제
