+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -2,34 +2,67 @@ package ladder.domain.ladder;
 
 
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import ladder.domain.participant.Participants;
 import ladder.domain.strategy.NextPointGenerationStrategy;
 
 public class Ladder {
 
-  private final Lines lines;
+  private final List<Line> lines;
 
-  private Ladder(Lines lines) {
+  private Ladder(List<Line> lines) {
     this.lines = lines;
   }
 
+  Ladder (Line... lines){
+    this(List.of(lines));
+  }
+
   public static Ladder createLadder(Height height, Participants participants, NextPointGenerationStrategy generationStrategy) {
-    Lines lines = Lines.createLines(height, participants, generationStrategy);
+    List<Line> lines = Stream.generate(() -> Line.createLine(participants.size(), generationStrategy))
+        .limit(height.height())
+        .collect(Collectors.toList());
     return new Ladder(lines);
   }
 
 
   public int height() {
-    return lines.height();
+    return lines.size();
   }
 
-  public List<Boolean> getRow(int rowNumber) {
-    return lines.getRow(rowNumber).getPoints();
+  private Line getRow(int rowNumber) {
+    return lines.get(rowNumber);
   }
+
+  public List<Boolean> getPointsOfRow(int rowNumber) {
+    return getRow(rowNumber).getPoints();
+  }
+
+
 
   public int getIndexOfResult(int startIndex) {
-    return lines.getLastIndex(startIndex);
+    int indexOfResult = startIndex;
+
+    for (int rowNumber = 0; rowNumber < height(); rowNumber++) {
+      indexOfResult = updateIndexOfResult(indexOfResult, rowNumber);
+    }
+
+    return indexOfResult;
   }
+
+  private int updateIndexOfResult(int indexOfResult, int rowNumber) {
+    if (getRow(rowNumber).isRightConnected(indexOfResult)) {
+      return indexOfResult + 1;
+    }
+
+    if (getRow(rowNumber).isLeftConnected(indexOfResult)) {
+      return indexOfResult - 1;
+    }
+
+    return indexOfResult;
+  }
+
 
 
 }
